requirements.txt
uvicorn
fastapi
python-dotenv
requests
openai
pydantic
